Изменение : Корректировка debug вывода.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@3556 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/filters/parser/MatroskaSplitter/MatroskaSplitter.cpp
@@ -733,7 +733,7 @@ HRESULT CMatroskaSplitterFilter::CreateOutputs(IAsyncReader* pAsyncReader)
 					DLog(L"CMatroskaSplitterFilter::CreateOutputs(): FrameRate = %.6f (by pTE->v.FrameRate)", framerate);
 				} else if (pTE->DefaultDuration > 0) {
 					framerate = 100.0 * UNITS / pTE->DefaultDuration;
-					DLog(L"CMatroskaSplitterFilter::CreateOutputs(): FrameRate = %.6f (by pTE->DefaultDuration = %I64d)", framerate, pTE->DefaultDuration);
+					DLog(L"CMatroskaSplitterFilter::CreateOutputs(): FrameRate = %.6f (by pTE->DefaultDuration = %I64d)", framerate, (INT64)pTE->DefaultDuration);
 				}
 
 				if (CodecID == "V_DSHOW/MPEG1VIDEO" || CodecID == "V_MPEG1") {
